requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
numpy
psycopg2-binary
gunicorn
django-heroku